Create independent component for CollectionConfig alert
